create-token pinocchio: address review feedback
- prepare.mjs: dump program via 'solana program dump -um' instead of
  mutating the user's Solana CLI config
- reuse pinocchio_token Mint::LEN instead of a hand-rolled MINT_SIZE const
- move CreateTokenArgs into create_token.rs (ix-specific) and the borsh
  string helpers into a util module; slim mod.rs to module wiring
- tests: run async bankrun setup in a before() hook so the it() blocks
  actually register (a describe callback runs synchronously)
